# src/frontend/openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "BatchRequest": {
        "properties": {
          "execute_orders": {
            "default": true,
            "description": "是否执行下单，false 时仅分析",
            "title": "Execute Orders",
            "type": "boolean"
          },
          "max_concurrency": {
            "default": 5,
            "description": "最大并发",
            "maximum": 20.0,
            "minimum": 1.0,
            "title": "Max Concurrency",
            "type": "integer"
          },
          "tickers": {
            "description": "股票代码列表",
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "title": "Tickers",
            "type": "array"
          }
        },
        "required": [
          "tickers"
        ],
        "title": "BatchRequest",
        "type": "object"
      },
      "CancelAllOrdersRequest": {
        "properties": {
          "reason": {
            "default": "manual",
            "description": "批量撤单原因",
            "title": "Reason",
            "type": "string"
          }
        },
        "title": "CancelAllOrdersRequest",
        "type": "object"
      },
      "ChannelSwitchRequest": {
        "properties": {
          "channel": {
            "description": "目标交易通道：simulation/ths_auto/ths_ipc/qmt",
            "title": "Channel",
            "type": "string"
          },
          "reconnect": {
            "default": true,
            "description": "切换后是否立即尝试连接新通道",
            "title": "Reconnect",
            "type": "boolean"
          }
        },
        "required": [
          "channel"
        ],
        "title": "ChannelSwitchRequest",
        "type": "object"
      },
...
      "DayRollRequest": {
        "properties": {
          "force": {
            "default": false,
            "description": "是否强制在非交易日执行",
            "title": "Force",
            "type": "boolean"
          },
          "reason": {
            "default": "manual",
            "description": "日切触发原因",
            "title": "Reason",
            "type": "string"
          }
        },
        "title": "DayRollRequest",
        "type": "object"
      },
      "DirectOrderRequest": {
        "properties": {
          "channel": {
            "default": "",
            "description": "交易通道，默认使用当前服务通道",
            "title": "Channel",
            "type": "string"
          },
          "client_order_id": {
            "default": "",
            "description": "外部系统订单号",
            "title": "Client Order Id",
            "type": "string"
          },
          "idempotency_key": {
            "description": "幂等键",
            "title": "Idempotency Key",
            "type": "string"
          },
          "manual_confirm": {
            "default": false,
            "description": "人工确认标记",
            "title": "Manual Confirm",
            "type": "boolean"
          },
          "manual_confirm_token": {
            "default": "",
            "description": "人工确认 token（可选）",
            "title": "Manual Confirm Token",
            "type": "string"
          },
          "memo": {
            "default": "",
            "description": "备注",
            "title": "Memo",
            "type": "string"
          },
          "order_type": {
            "default": "",
            "description": "limit 或 market",
            "title": "Order Type",
            "type": "string"
          },
          "price": {
            "description": "委托价格",
            "exclusiveMinimum": 0.0,
            "title": "Price",
            "type": "number"
          },
          "qty": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "委托数量（兼容字段）",
            "title": "Qty"
          },
          "quantity": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "委托数量",
            "title": "Quantity"
          },
          "request_id": {
            "default": "",
            "description": "外部请求 ID（可选）",
            "title": "Request Id",
            "type": "string"
          },
          "side": {
            "description": "BUY 或 SELL",
            "title": "Side",
            "type": "string"
          },
          "ticker": {
            "description": "股票代码，例如 000001",
            "title": "Ticker",
            "type": "string"
          },
          "trace_id": {
            "default": "",
            "description": "全链路追踪 ID（可选）",
            "title": "Trace Id",
            "type": "string"
          },
          "type": {
            "default": "",
            "description": "limit 或 market（兼容字段）",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "ticker",
          "side",
          "price",
          "idempotency_key"
        ],
        "title": "DirectOrderRequest",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ReconcileRequest": {
        "properties": {
          "initial_cash": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "可选，对账初始现金基准",
            "title": "Initial Cash"
          }
        },
        "title": "ReconcileRequest",
        "type": "object"
      },
...
      "TickerRequest": {
        "properties": {
          "ticker": {
            "description": "股票代码，例如 000001",
            "title": "Ticker",
            "type": "string"
          }
        },
        "required": [
          "ticker"
        ],
        "title": "TickerRequest",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "AI 驱动量化交易客户端",
    "title": "来财 (Attract-wealth)",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/trading/analyze": {
      "post": {
        "operationId": "analyze_api_trading_analyze_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TickerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Analyze",
        "tags": [
          "trading"
        ]
      }
    },
...
    "/api/trading/batch": {
      "post": {
        "operationId": "batch_api_trading_batch_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Batch",
        "tags": [
          "trading"
        ]
      }
    },
    "/api/trading/channel/switch": {
      "post": {
        "description": "切换交易通道并重建执行上下文。",
        "operationId": "switch_trading_channel_api_trading_channel_switch_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChannelSwitchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Switch Trading Channel",
        "tags": [
          "trading"
        ]
      }
    },
    "/api/trading/day-roll": {
      "post": {
        "operationId": "day_roll_api_trading_day_roll_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DayRollRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Day Roll",
        "tags": [
          "trading"
        ]
      }
    },
    "/api/trading/execute": {
      "post": {
        "operationId": "execute_api_trading_execute_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TickerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Execute",
        "tags": [
          "trading"
        ]
      }
    },
...
    "/api/trading/orders/cancel-all": {
      "post": {
        "description": "批量撤销当前活动订单。",
        "operationId": "cancel_all_orders_api_trading_orders_cancel_all_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CancelAllOrdersRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel All Orders",
        "tags": [
          "trading"
        ]
      }
    },
    "/api/trading/orders/direct": {
      "post": {
        "operationId": "direct_order_api_trading_orders_direct_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DirectOrderRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Direct Order",
        "tags": [
          "trading"
        ]
      }
    },
...
    "/api/trading/orders/trace": {
      "get": {
        "operationId": "order_trace_api_trading_orders_trace_get",
        "parameters": [
          {
            "in": "query",
            "name": "idempotency_key",
            "required": false,
            "schema": {
              "default": "",
              "title": "Idempotency Key",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "local_order_id",
            "required": false,
            "schema": {
              "default": "",
              "title": "Local Order Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "client_order_id",
            "required": false,
            "schema": {
              "default": "",
              "title": "Client Order Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "request_id",
            "required": false,
            "schema": {
              "default": "",
              "title": "Request Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "trace_id",
            "required": false,
            "schema": {
              "default": "",
              "title": "Trace Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Order Trace",
        "tags": [
          "trading"
        ]
      }
    },
...
    "/api/trading/reconcile": {
      "post": {
        "operationId": "reconcile_api_trading_reconcile_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReconcileRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reconcile",
        "tags": [
          "trading"
        ]
      }
    },
    "/api/trading/reconcile/unlock": {
      "post": {
        "operationId": "reconcile_unlock_api_trading_reconcile_unlock_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReconcileUnlockRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reconcile Unlock",
        "tags": [
          "trading"
        ]
      }
    },
    "/api/trading/snapshot": {
      "get": {
        "operationId": "trade_snapshot_api_trading_snapshot_get",
        "parameters": [
          {
            "description": "是否包含通道原始快照",
            "in": "query",
            "name": "include_channel_raw",
            "required": false,
            "schema": {
              "default": true,
              "description": "是否包含通道原始快照",
              "title": "Include Channel Raw",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Trade Snapshot",
        "tags": [
          "trading"
        ]
      }
    },
...
  }
}